:koko: Split esClient from functions
